scheduler/requirements.txt
anthropic>=0.39.0
slack_sdk>=3.27.0
schedule>=1.2.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
